requirements.txt
affine==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
awscli==1.38.35
AWSIoTPythonSDK==1.5.4
bidict==0.23.1
boto3==1.37.35
botocore==1.37.35
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.2.0
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docutils==0.19
et_xmlfile==2.0.0
fastapi==0.115.14
fonttools==4.58.0
frozenlist==1.7.0
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.6.3
nicegui==2.20.0
numpy==2.2.6
oauthlib==3.2.2
openpyxl==3.1.5
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.2
pscript==0.7.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
rasterio==1.4.3
requests==2.32.4
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.11.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
tweepy==4.14.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.35.0
vbuild==0.8.2
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
python-dotenv==1.1.1
